server/requirements.txt
aniso8601==9.0.1
async-timeout==4.0.2
bertopic==0.6.0
bidict==0.21.4
certifi==2022.12.7
charset-normalizer==2.0.12
clean-text==0.5.0
click==8.0.4
Cython==0.29.32
dataclasses==0.8
dnspython==1.16.0
emoji==1.7.0
eventlet==0.30.2
filelock==3.4.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JSON==0.3.5
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.2
ftfy==6.0.3
greenlet==2.0.1
gunicorn==20.1.0
hdbscan==0.8.27
huggingface-hub==0.4.0
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.1
llvmlite==0.36.0
MarkupSafe==2.0.1
nltk==3.6.7
numba==0.53.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pynndescent==0.5.8
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
PyYAML==6.0
redis==4.3.5
regex==2022.10.31
requests==2.27.1
rq==1.11.1
sacremoses==0.0.53
scikit-learn==0.24.2
scipy==1.5.4
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
transformers==4.18.0
typing_extensions==4.1.1
umap-learn==0.5.3
Unidecode==1.3.6
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.0.3
zipp==3.6.0
